Platform-specific networking constants
